% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clews.R
\name{add_modelled_npp_to_zhang_npp_measurements}
\alias{add_modelled_npp_to_zhang_npp_measurements}
\title{Add modelled NPP to measured zhang data}
\usage{
add_modelled_npp_to_zhang_npp_measurements(zhang_data, modelled_npp)
}
\arguments{
\item{zhang_data}{zhang data list object 
(e.g. obtained from read_in_zhang_file() function)}

\item{modelled_npp}{array with modelled NPP (dimensions [lat,lon,year])}
}
\value{
modified zhang_data input list, with added colums: measured, modelled
}
\description{
Add the best fitting modelled NPP to the measured zhang data list object 
(e.g. obtained from read_in_zhang_file() function)
}
\examples{
\dontrun{
}

}
