% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clews.R
\name{climate_plot}
\alias{climate_plot}
\title{Plot the climate of a given location}
\usage{
climate_plot(temp, prec, file = NULL, title = "")
}
\arguments{
\item{file}{character path. file location to save the plot to. 
If not supplied prints to screen. Default: NULL}

\item{title}{character string printed as title.}

\item{tmp}{temperature array with 12 values for each month in °C}

\item{pre}{precipitation array with 12 values for each month in mm}
}
\value{
None
}
\description{
Plot an array of lon_lat data to screen or file 
e.g. obtained from read_netcdf and averaged over time.
Prints to screen if file argument is not supplied (default).
}
\examples{
\dontrun{
}

}
