% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clews.R
\name{get_main_variable}
\alias{get_main_variable}
\title{Get main variable of a netcdf file}
\usage{
get_main_variable(file_nc)
}
\arguments{
\item{file_nc}{netcdf file name}
}
\value{
character string: determined main variable name
}
\description{
Utility function to guess the main variable of a netcdf file
}
\examples{
\dontrun{
}

}
