% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clews.R
\name{get_netcdf_cellsize}
\alias{get_netcdf_cellsize}
\title{Returns cellarea of given netcdf in m}
\usage{
get_netcdf_cellsize(nc_in_file = tmp_file)
}
\arguments{
\item{nc_in_file}{netcdf file name}
}
\value{
cellarea array with dimensions [longitude, latitude] in m
}
\description{
Returns approximate cellarea of given netcdf file in m
}
\examples{
\dontrun{
}

}
