% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clews.R
\name{get_timestep}
\alias{get_timestep}
\title{Get time resolution of a netcdf file}
\usage{
get_timestep(file_nc)
}
\arguments{
\item{file_nc}{netcdf file name}
}
\value{
character string: determined timestep (daily, monthly, yearly)
}
\description{
Utility function to determine time resolution of a netcdf file
}
\examples{
\dontrun{
}

}
