% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clews.R
\name{plot_lon_lat_array}
\alias{plot_lon_lat_array}
\title{Plot an array of lon_lat data to screen or file}
\usage{
plot_lon_lat_array(data, file = NULL, title = "")
}
\arguments{
\item{data}{array to plot. Needs to be in format [longitude,latitude].}

\item{file}{character path. file location to save the plot to. 
If not supplied prints to screen. Default: NULL}

\item{title}{character string printed as title.}
}
\value{
None
}
\description{
Plot an array of lon_lat data to screen or file 
e.g. obtained from read_netcdf and averaged over time.
Prints to screen if file argument is not supplied (default).
}
\examples{
\dontrun{
}

}
