% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clews.R
\name{read_in_zhang_file}
\alias{read_in_zhang_file}
\title{Read in the Zhang et al. csv file with NPP measurements}
\usage{
read_in_zhang_file(csv_in_file)
}
\arguments{
\item{csv_in_file}{character string. full csv file path}
}
\value{
list object with dimensions [lat, lon, totnpp, year] per record
}
\description{
Read in the Zhang et al. csv file with NPP measurements and return a 
list of relevant records
}
\examples{
\dontrun{
}

}
