% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clews.R
\name{read_netcdf}
\alias{read_netcdf}
\title{Reads netcdf and returns the requested variable as an array}
\usage{
read_netcdf(
  nc_in_file,
  var = NULL,
  get_year_start = NULL,
  get_year_stop = NULL,
  suppress_read_print = FALSE
)
}
\arguments{
\item{nc_in_file}{netcdf file name}

\item{var}{optional variable to be read, in case automatic detection does 
not work as intended or several variables are stored within the file}

\item{get_year_start}{first year to be read 
(if not specified will default to first record year)}

\item{get_year_stop}{final year to be read 
(if not specified will default to last record year)}

\item{suppress_read_print}{whether to suppress the info "reading file XYZ" 
(default FALSE)}
}
\value{
array with netcdf's data, dim=c([longitude],[latitude],[bands],[months],[years])
}
\description{
Reads an arbitrary netcdf and returns the requested variable in the given 
year range as an array
}
\examples{
\dontrun{
}

}
