% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clews.R
\name{scatter_plot}
\alias{scatter_plot}
\title{Create a scatterplot}
\usage{
scatter_plot(x, y, file = NULL, title = "", xlab = "", ylab = "")
}
\arguments{
\item{x}{data to plot on x axis}

\item{y}{data to plot on y axis}

\item{file}{character path. file location to save the plot to. 
If not supplied prints to screen. Default: NULL}

\item{title}{character string printed as title.}

\item{xlab}{label for x axis}

\item{ylab}{label for y axis}
}
\value{
None
}
\description{
Create a scatterplot of two data series e.g. measured vs. modelled
}
\examples{
\dontrun{
}

}
